(feat): pantallas de registro y login implementadas (sin logica aun)
