+++ package.json
@@ -16,6 +16,7 @@
     "react-dom": "^19.1.1",
     "react-icons": "^5.5.0",
     "react-markdown": "^10.1.0",
+    "react-router-dom": "^7.12.0",
     "react-syntax-highlighter": "^16.1.0"
   },
   "devDependencies": {
