+++ src/App.jsx
@@ -1,381 +1,37 @@
-import { useState, useEffect, useRef } from 'react'
+import { useState, useEffect } from 'react'
+import { Routes, Route } from 'react-router-dom'
 import './css/App.css'
-import ChatMessage from './components/ChatMessage'
-import ChatInput from './components/ChatInput'
-import ThemeToggle from './components/ThemeToggle'
-import CodeSidebar from './components/CodeSidebar'
-import LeftSidebar from './components/LeftSidebar'
+import Chat from './components/Chat'
+import Login from './components/Login'
+import Register from './components/Register'
 
 function App() {
-  const [messages, setMessages] = useState([])
-  const [isLoading, setIsLoading] = useState(false)
-  const [selectedModel, setSelectedModel] = useState('')
-  const [images, setImages] = useState([])
-  const [isDarkTheme, setIsDarkTheme] = useState(() => {
-    // Detectar preferencia del sistema
-    return window.matchMedia('(prefers-color-scheme: dark)').matches
-  })
-  const [codeRequests, setCodeRequests] = useState([])
-  const [isLeftSidebarOpen, setIsLeftSidebarOpen] = useState(true)
-  const messagesEndRef = useRef(null)
-  const autoScrollEnabled = useRef(true)
-  const messagesContainerRef = useRef(null)
-  const isAutoScrolling = useRef(false)
+  const [isAuthenticated, setIsAuthenticated] = useState(false)
 
-  // Aplicar tema
+  // Verificar si el usuario está autenticado al cargar la app
   useEffect(() => {
-    document.documentElement.setAttribute('data-theme', isDarkTheme ? 'dark' : 'light')
-  }, [isDarkTheme])
+    const authStatus = localStorage.getItem('isAuthenticated')
+    setIsAuthenticated(authStatus === 'true')
 
-  // Extraer bloques de código de los mensajes de la IA
-  useEffect(() => {
-    const extractCodeBlocks = () => {
-      const requests = []
-      let currentRequestCodes = []
-      let currentUserMessage = ''
-      let isCollectingCodes = false
-
-      messages.forEach((msg, index) => {
-        // Detectar inicio de una nueva petición (mensaje del usuario)
-        if (msg.isUser) {
-          // Si había códigos de la petición anterior, guardarlos
-          if (currentRequestCodes.length > 0) {
-            requests.push({ 
-              codes: currentRequestCodes,
-              userMessage: currentUserMessage
-            })
-            currentRequestCodes = []
-          }
-          isCollectingCodes = true
-          currentUserMessage = msg.text
-        } 
-        // Extraer códigos de respuestas de la IA
-        else if (isCollectingCodes && !msg.isUser && !msg.isError) {
-          const codeBlockPattern = /```(\w+)?\n([\s\S]*?)```/g
-          let match
-          
-          while ((match = codeBlockPattern.exec(msg.text)) !== null) {
-            const language = match[1] || 'text'
-            const content = match[2].trim()
-            
-            if (content) {
-              currentRequestCodes.push({
-                content,
-                language
-              })
-            }
-          }
-        }
-      })
-
-      // Guardar los códigos de la última petición si existen
-      if (currentRequestCodes.length > 0) {
-        requests.push({ 
-          codes: currentRequestCodes,
-          userMessage: currentUserMessage
-        })
-      }
-
-      setCodeRequests(requests)
+    // Escuchar eventos personalizados de cambio de autenticación
+    const handleAuthChange = () => {
+      const authStatus = localStorage.getItem('isAuthenticated')
+      setIsAuthenticated(authStatus === 'true')
     }
 
-    extractCodeBlocks()
-  }, [messages])
+    window.addEventListener('authChange', handleAuthChange)
 
-  // Detectar scroll manual del usuario
-  useEffect(() => {
-    const container = messagesContainerRef.current
-    if (!container) return
-
-    let scrollTimeout = null
-    let lastScrollTop = container.scrollTop
-
-    const handleScroll = () => {
-      // Ignorar eventos de scroll si estamos haciendo autoscroll
-      if (isAutoScrolling.current) {
-        lastScrollTop = container.scrollTop
-        return
-      }
-
-      const currentScrollTop = container.scrollTop
-      
-      // Solo procesar si hay un cambio real de posición (scroll manual)
-      if (currentScrollTop === lastScrollTop) {
-        return
-      }
-      
-      lastScrollTop = currentScrollTop
-
-      // Limpiar timeout anterior
-      if (scrollTimeout) {
-        clearTimeout(scrollTimeout)
-      }
-
-      // Esperar un poco para asegurarnos de que el usuario terminó de scrollear
-      scrollTimeout = setTimeout(() => {
-        const isAtBottom = container.scrollHeight - container.scrollTop - container.clientHeight < 10
-        
-        // Si NO estamos al final, desactivar autoscroll
-        if (!isAtBottom) {
-          autoScrollEnabled.current = false
-        } else {
-          // Si volvemos al final manualmente, reactivar autoscroll
-          autoScrollEnabled.current = true
-        }
-      }, 50)
-    }
-
-    container.addEventListener('scroll', handleScroll, { passive: true })
     return () => {
-      container.removeEventListener('scroll', handleScroll)
-      if (scrollTimeout) clearTimeout(scrollTimeout)
+      window.removeEventListener('authChange', handleAuthChange)
     }
   }, [])
 
-  // Auto-scroll cuando está habilitado y hay cambios en los mensajes
-  useEffect(() => {
-    if (autoScrollEnabled.current && messagesEndRef.current) {
-      isAutoScrolling.current = true
-      messagesEndRef.current.scrollIntoView({ behavior: 'smooth' })
-      
-      // Resetear la bandera después de que termine el scroll
-      setTimeout(() => {
-        isAutoScrolling.current = false
-      }, 1000)
-    }
-  }, [messages])
-
-  const handleSendMessage = async (userMessage) => {
-    // Activar auto-scroll cuando el usuario envía un mensaje
-    autoScrollEnabled.current = true
-    
-    // Determinar si el modelo seleccionado tiene capacidades de visión
-    const isVisionModel = () => {
-      if (!selectedModel || selectedModel === 'No hay LLMs') return false
-      if (selectedModel === 'Auto') return true
-      
-      const modelLower = selectedModel.toLowerCase()
-      const visionKeywords = ['vl', 'vision', 'llava', 'bakllava', 'moondream']
-      return visionKeywords.some(keyword => modelLower.includes(keyword))
-    }
-    
-    // Filtrar imágenes: solo incluir si el modelo tiene visión
-    const imagesToSend = isVisionModel() ? [...images] : []
-    
-    // Agregar mensaje del usuario con imágenes solo si el modelo tiene visión
-    const newUserMessage = { 
-      text: userMessage, 
-      isUser: true,
-      images: imagesToSend
-    }
-    setMessages(prev => [...prev, newUserMessage])
-    
-    // Limpiar imágenes al enviar
-    setImages([])
-    
-    // Agregar mensaje vacío de la IA que se irá llenando
-    const aiMessageIndex = messages.length + 1
-    setMessages(prev => [...prev, { text: '', isUser: false, isLoading: true }])
-    setIsLoading(true)
-
-    try {
-      // Determinar el modelo a usar
-      const modelToUse = selectedModel === 'Auto' ? 'qwen2.5-coder:14b' : selectedModel
-      const isAutoMode = selectedModel === 'Auto'
-      
-      // Construir historial de mensajes para contexto (incluyendo el mensaje actual)
-      const messageHistory = [...messages, newUserMessage].map(msg => ({
-        role: msg.isUser ? 'user' : 'assistant',
-        content: msg.text
-      }))
-      
-      // Crear FormData para enviar al backend
-      const formData = new FormData()
-      formData.append('model', modelToUse)
-      formData.append('prompt', userMessage)
-      formData.append('messages', JSON.stringify(messageHistory))
-      formData.append('autoMode', isAutoMode ? 'true' : 'false')
-      
-      // Añadir imágenes solo si el modelo tiene visión
-      if (imagesToSend.length > 0) {
-        imagesToSend.forEach((img) => {
-          formData.append('images', img.file)
-        })
-      }
-
-      // Llamar al backend con streaming
-      const response = await fetch('http://localhost:3000/api/generate/stream', {
-        method: 'POST',
-        body: formData,
-      })
-
-      if (!response.ok) {
-        throw new Error('Error en la respuesta del servidor')
-      }
-
-      // Leer el stream
-      const reader = response.body.getReader()
-      const decoder = new TextDecoder()
-      let accumulatedText = ''
-      let step1Text = ''
-      let step2Text = ''
-      let currentStep = 0 // 0: ninguno, 1: paso 1 (PlantUML), 2: paso 2 (respuesta final)
-
-      while (true) {
-        const { done, value } = await reader.read()
-        if (done) break
-
-        const chunk = decoder.decode(value, { stream: true })
-        const lines = chunk.split('\n')
-        
-        for (const line of lines) {
-          if (line.startsWith('data: ')) {
-            const data = line.slice(6)
-            if (data === '[DONE]') {
-              break
-            }
-            if (data.startsWith('[ERROR]')) {
-              throw new Error(data.slice(8))
-            }
-            
-            // Detectar eventos de control de pasos
-            try {
-              const decodedChunk = JSON.parse(data)
-              
-              if (decodedChunk === '[STEP1_START]') {
-                currentStep = 1
-                continue
-              }
-              
-              if (decodedChunk === '[STEP1_END]') {
-                step1Text = accumulatedText
-                accumulatedText = ''
-                continue
-              }
-              
-              if (decodedChunk === '[STEP2_START]') {
-                currentStep = 2
-                continue
-              }
-              
-              // Acumular texto según el paso actual
-              if (currentStep === 1) {
-                accumulatedText += decodedChunk
-                // Actualizar el mensaje con el paso 1 en progreso
-                setMessages(prev => {
-                  const newMessages = [...prev]
-                  newMessages[aiMessageIndex] = { 
-                    text: accumulatedText, 
-                    isUser: false, 
-                    isLoading: false,
-                    isTwoStep: true,
-                    step1Text: accumulatedText,
-                    step2Text: '',
-                    currentStep: 1
-                  }
-                  return newMessages
-                })
-              } else if (currentStep === 2) {
-                accumulatedText += decodedChunk
-                // Actualizar el mensaje con ambos pasos
-                setMessages(prev => {
-                  const newMessages = [...prev]
-                  newMessages[aiMessageIndex] = { 
-                    text: accumulatedText, 
-                    isUser: false, 
-                    isLoading: false,
-                    isTwoStep: true,
-                    step1Text: step1Text,
-                    step2Text: accumulatedText,
-                    currentStep: 2
-                  }
-                  return newMessages
-                })
-              } else {
-                // Modo normal (sin pasos)
-                accumulatedText += decodedChunk
-                setMessages(prev => {
-                  const newMessages = [...prev]
-                  newMessages[aiMessageIndex] = { text: accumulatedText, isUser: false, isLoading: false }
-                  return newMessages
-                })
-              }
-            } catch (e) {
-              // Si no es JSON válido, tratar como texto normal
-              accumulatedText += data
-              setMessages(prev => {
-                const newMessages = [...prev]
-                newMessages[aiMessageIndex] = { text: accumulatedText, isUser: false, isLoading: false }
-                return newMessages
-              })
-            }
-          }
-        }
-      }
-      
-    } catch (error) {
-      console.error('Error al enviar mensaje:', error)
-      // Actualizar el mensaje de la IA con el error
-      setMessages(prev => {
-        const newMessages = [...prev]
-        newMessages[aiMessageIndex] = {
-          text: 'Error al comunicarse con la IA. Verifica que el backend esté corriendo.',
-          isUser: false,
-          isError: true
-        }
-        return newMessages
-      })
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
-  const toggleTheme = () => {
-    setIsDarkTheme(prev => !prev)
-  }
-
   return (
-    <div className={`app-container ${isLeftSidebarOpen ? 'left-sidebar-open' : ''}`}>
-      <ThemeToggle isDark={isDarkTheme} onToggle={toggleTheme} />
-      
-      <LeftSidebar isOpen={isLeftSidebarOpen} setIsOpen={setIsLeftSidebarOpen} />
-      
-      <CodeSidebar codeRequests={codeRequests} />
-
-      <div className="messages-container" ref={messagesContainerRef}>
-        {messages.length === 0 ? (
-          <div className="empty-state">
-            <p>¿En qué puedo ayudarte hoy?</p>
-          </div>
-        ) : (
-          messages.map((msg, index) => (
-            <ChatMessage 
-              key={index} 
-              message={msg.text} 
-              isUser={msg.isUser}
-              isError={msg.isError}
-              images={msg.images || []}
-              isFirstMessage={index === 0}
-              isLoading={msg.isLoading || false}
-              isTwoStep={msg.isTwoStep || false}
-              step1Text={msg.step1Text || ''}
-              step2Text={msg.step2Text || ''}
-              currentStep={msg.currentStep || 0}
-            />
-          ))
-        )}
-        <div ref={messagesEndRef} />
-      </div>
-
-      <ChatInput 
-        onSendMessage={handleSendMessage} 
-        isLoading={isLoading}
-        selectedModel={selectedModel}
-        onModelChange={setSelectedModel}
-        images={images}
-        onImagesChange={setImages}
-      />
-    </div>
+    <Routes>
+      <Route path="/" element={<Chat isAuthenticated={isAuthenticated} />} />
+      <Route path="/login" element={<Login />} />
+      <Route path="/register" element={<Register />} />
+    </Routes>
   )
 }
