+++ src/css/Login.css
@@ -1,0 +1,235 @@
+.auth-container {
+  display: flex;
+  min-height: 100vh;
+  background-color: #000000;
+  overflow: hidden;
+  position: relative;
+}
+
+/* Logo pequeño arriba a la izquierda */
+.top-logo {
+  position: absolute;
+  top: 2rem;
+  left: 2rem;
+  z-index: 10;
+  display: none;
+}
+
+.top-logo img {
+  height: 20px;
+  width: auto;
+}
+
+/* Mitad izquierda - Formulario */
+.auth-left {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 2rem;
+  z-index: 2;
+}
+
+.auth-card {
+  width: 100%;
+  max-width: 450px;
+}
+
+/* Mitad derecha - Logo */
+.auth-right {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  overflow: hidden;
+}
+
+.logo-container {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: 100%;
+}
+
+
+.logo-large {
+  width: 400px;
+  height: auto;
+  position: relative;
+  z-index: 1;
+  filter: drop-shadow(0 0 30px rgba(125, 57, 235, 0.3));
+}
+
+.auth-logo {
+  display: flex;
+  justify-content: center;
+  margin-bottom: 2rem;
+}
+
+.auth-logo img {
+  width: 200px;
+  height: auto;
+}
+
+.auth-title {
+  font-size: 2.5rem;
+  font-weight: 600;
+  color: #ffffff;
+  text-align: center;
+  margin: 0 0 4rem 0;
+}
+
+.auth-subtitle {
+  font-size: 1rem;
+  color: var(--text-color);
+  opacity: 0.7;
+  text-align: center;
+  margin: 0 0 2rem 0;
+}
+
+.auth-form {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
+
+.form-group {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+
+.form-group label {
+  font-size: 0.95rem;
+  font-weight: 500;
+  color: #ffffff;
+}
+
+.password-input-wrapper {
+  position: relative;
+  display: flex;
+  align-items: center;
+}
+
+.form-group input {
+  padding: 0.75rem 1rem;
+  font-size: 1rem;
+  border: 1px solid #333333;
+  border-radius: 0.5rem;
+  background-color: #1a1a1a;
+  color: #ffffff;
+  transition: all 0.2s ease;
+  width: 100%;
+}
+
+.password-input-wrapper input {
+  padding-right: 2.75rem;
+}
+
+.password-toggle {
+  position: absolute;
+  right: 0.75rem;
+  background: none;
+  border: none;
+  color: #7D39EB;
+  cursor: pointer;
+  padding: 0.5rem;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 1.25rem;
+  transition: opacity 0.2s ease;
+}
+
+.password-toggle:hover {
+  opacity: 0.7;
+}
+
+.form-group input:focus {
+  outline: none;
+  border-color: #7D39EB;
+  box-shadow: 0 0 0 3px rgba(125, 57, 235, 0.2);
+}
+
+.form-group input::placeholder {
+  color: #666666;
+  opacity: 1;
+}
+
+.auth-button {
+  margin-top: 1rem;
+  padding: 0.875rem 1rem;
+  font-size: 1rem;
+  font-weight: 600;
+  color: white;
+  background-color: #7D39EB;
+  border: none;
+  border-radius: 0.5rem;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.auth-button:hover {
+  background-color: #6B2FD1;
+  transform: translateY(-1px);
+  box-shadow: 0 4px 12px rgba(125, 57, 235, 0.3);
+}
+
+.auth-button:active {
+  transform: translateY(0);
+}
+
+.auth-footer {
+  margin-top: 4rem;
+  text-align: center;
+  font-size: 1rem;
+  color: #999999;
+}
+
+.auth-link {
+  color: #7D39EB;
+  font-weight: 600;
+  text-decoration: none;
+  transition: opacity 0.2s ease;
+}
+
+.auth-link:hover {
+  opacity: 0.8;
+  text-decoration: underline;
+}
+
+/* Responsive */
+@media (max-width: 1024px) {
+  .auth-container {
+    flex-direction: column;
+  }
+
+  .auth-right {
+    display: none;
+  }
+
+  .auth-left {
+    flex: 1;
+  }
+
+  .top-logo {
+    display: block;
+  }
+}
+
+@media (max-width: 768px) {
+  .auth-card {
+    padding: 2rem;
+  }
+
+  .auth-title {
+    font-size: 1.75rem;
+  }
+
+  .logo-large {
+    width: 250px;
+  }
+}
